building the main

+++ src/vehicle/Motorized.java
@@ -4,7 +4,7 @@ public interface Motorized {
     /**
      * This interface is used to define the methods that are used in the Motorized class
      */
-    public void setAverageFuelConsumption(float fuelConsumption);
+    public void setAverageFuelConsumption(int fuelConsumption);
     public float getAverageFuelConsumption();
 
     public void setAverageEngineLife(int engineLife);
